CORE-5995 REFACTOR: reuse duplicates in injector code
