+++ components/flow/flow-service/src/integrationTest/kotlin/net/corda/flow/testing/fakes/FakeSandboxGroupContextComponent.kt
@@ -151,7 +151,7 @@ class FakeSandboxGroupContextComponent : SandboxGroupContextComponent {
         override fun injectServices(flow: Flow) {
         }
 
-        override fun getRegisteredServices(): Set<SingletonSerializeAsToken> {
+        fun getRegisteredServices(): Set<SingletonSerializeAsToken> {
             return emptySet()
         }
