+++ components/virtual-node/sandbox-group-context-service/src/main/java/net/corda/sandboxgroupcontext/service/impl/package-info.java
@@ -1,0 +1,4 @@
+@Export
+package net.corda.sandboxgroupcontext.service.impl;
+
+import org.osgi.annotation.bundle.Export;
