+++ components/virtual-node/sandbox-group-context-service/src/main/kotlin/net/corda/sandboxgroupcontext/service/SandboxDependencyInjector.kt
@@ -1,0 +1,10 @@
+package net.corda.sandboxgroupcontext.service
+
+import net.corda.v5.serialization.SingletonSerializeAsToken
+
+interface SandboxDependencyInjector: AutoCloseable {
+    /**
+     * @return A collection of services registered with the injector.
+     */
+    fun getRegisteredServices(): Collection<SingletonSerializeAsToken>
+}
